+++ udaddons/updraftplus-addons.php
@@ -26,6 +26,8 @@ $udaddons2_mothership = (defined('UPDRAFTPLUS_ADDONS_SSL') && !UPDRAFTPLUS_ADDON
 
 $udaddons2_mothership .= defined('UDADDONS2_TEST_MOTHERSHIP') ? UDADDONS2_TEST_MOTHERSHIP : 'updraftplus.com';
 
+global $updraftplus_addons2; // Need to explicitly globalise the variable or WP-CLI won't recognise it https://github.com/wp-cli/wp-cli/issues/4019#issuecomment-297410839
+
 $updraftplus_addons2 = new UpdraftPlusAddons2('updraftplus', $udaddons2_mothership);
 
 class UpdraftPlusAddons2 {
@@ -82,34 +84,29 @@ class UpdraftPlusAddons2 {
 		add_filter('site_transient_update_plugins', array($this, 'site_transient_update_plugins'), 9);
 
 		// Over-ride update mechanism for the plugin
-		if (is_readable(UPDRAFTPLUS_DIR.'/___disable_autoupdate___/vendor/yahnis-elsts/plugin-update-checker/plugin-update-checker.php')) {
+		if (is_readable(UPDRAFTPLUS_DIR.'/vendor/yahnis-elsts/plugin-update-checker/plugin-update-checker.php')) {
 
 			include_once(UPDRAFTPLUS_DIR.'/vendor/yahnis-elsts/plugin-update-checker/plugin-update-checker.php');
 
-			$options = $this->get_option(UDADDONS2_SLUG.'_options');
-			$email = isset($options['email']) ? $options['email'] : '';
-			if ($email) {
+			add_filter('puc_check_now-'.$this->slug, array($this, 'puc_check_now'), 10, 3);
+			add_filter('puc_retain_fields-'.$this->slug, array($this, 'puc_retain_fields'));
+			add_filter('puc_request_info_options-'.$this->slug, array($this, 'puc_request_info_options'));
+			// Run after the PluginUpdateChecker has done its stuff
+			add_filter('site_transient_update_plugins', array($this, 'possibly_inject_translations'), 11);
+			// If https does not work, then try http
 
-				add_filter('puc_check_now-'.$this->slug, array($this, 'puc_check_now'), 10, 3);
-				add_filter('puc_retain_fields-'.$this->slug, array($this, 'puc_retain_fields'));
-				add_filter('puc_request_info_options-'.$this->slug, array($this, 'puc_request_info_options'));
-				// Run after the PluginUpdateChecker has done its stuff
-				add_filter('site_transient_update_plugins', array($this, 'possibly_inject_translations'), 11);
-				// If https does not work, then try http
-
-				$plug_updatechecker = Puc_v4_Factory::buildUpdateChecker($this->url."/plugin-info/", WP_PLUGIN_DIR.'/'.$this->slug.'/'.$this->slug.'.php', $this->slug, 24);
+			$plug_updatechecker = Puc_v4_Factory::buildUpdateChecker($this->url."/plugin-info/", WP_PLUGIN_DIR.'/'.$this->slug.'/'.$this->slug.'.php', $this->slug, 24);
+			
+			// The null case is seen in HS#36382
+			if (null === $plug_updatechecker) {
+				error_log("UpdraftPlus: Puc_v4_Factory::buildUpdateChecker() return a null object");
+			} else {
+				$plug_updatechecker->addQueryArgFilter(array($this, 'updater_queryargs_plugin'));
+				if ($this->debug) $plug_updatechecker->debugMode = true;
 				
-				// The null case is seen in HS#36382
-				if (null === $plug_updatechecker) {
-					error_log("UpdraftPlus: Puc_v4_Factory::buildUpdateChecker() return a null object");
-				} else {
-					$plug_updatechecker->addQueryArgFilter(array($this, 'updater_queryargs_plugin'));
-					if ($this->debug) $plug_updatechecker->debugMode = true;
-					
-					$plug_updatechecker->addFilter('request_metadata_http_result', array($this, 'request_metadata_http_result'), 10, 3);
-					
-					$this->plug_updatechecker = $plug_updatechecker;
-				}
+				$plug_updatechecker->addFilter('request_metadata_http_result', array($this, 'request_metadata_http_result'), 10, 3);
+				
+				$this->plug_updatechecker = $plug_updatechecker;
 			}
 
 			$this->move_updraftplus_update_cron();
@@ -123,9 +120,9 @@ class UpdraftPlusAddons2 {
 	 */
 	private function move_updraftplus_update_cron() {
 
-		$cron_move_date = 1588896000; // 2020-05-08 00:00
-		$cleanup_date = 1588982400; // 2020-05-09 00:00
-		$option_name = 'updraftplus_move_update_cron_may20';
+		$cron_move_date = 1607558400; // 2020-12-10 00:00
+		$cleanup_date = 1607644800; // 2020-12-11 00:00
+		$option_name = 'updraftplus_move_update_cron_dec20';
 		
 		$time_now = time();
 		$start_of_today = $time_now - ($time_now % 86400);
@@ -136,18 +133,18 @@ class UpdraftPlusAddons2 {
 
 		if ($cron_move_date == $start_of_today && !$moved_cron) {
 			$timestamp = wp_next_scheduled('puc_cron_check_updates-updraftplus');
-			$begintime1 = 1588942800; // 2020-05-08 13:00
-			$endtime1 = $begintime1 + 3600;
-			$begintime2 = 1588953600; // 2020-05-08 16:00
-			$endtime2 = $begintime2 + 7 * 3600;
+			$begintime1 = $cron_move_date + 3600 * 9; // 09:00
+			$endtime1 = $begintime1 + 3600 * 12;
+			// $begintime2 = 1588953600; // 2020-05-08 16:00
+			// $endtime2 = $begintime2 + 7 * 3600;
 
-			if (($timestamp >= $begintime1 && $timestamp <= $endtime1) || ($timestamp >= $begintime2 && $timestamp <= $endtime2)) {
+			if ($timestamp >= $begintime1 && $timestamp <= $endtime1) {
 				wp_clear_scheduled_hook('puc_cron_check_updates-updraftplus');
-				wp_schedule_event($timestamp + 12 * 3600, 'daily', 'puc_cron_check_updates-updraftplus');
+				wp_schedule_event($timestamp + 13 * 3600, 'daily', 'puc_cron_check_updates-updraftplus');
 				$this->update_option($option_name, true);
 			}
 			
-		} elseif ($moved_cron) {
+		} elseif ($moved_cron && $start_of_today >= $cleanup_date) {
 			delete_site_option($option_name);
 		}
 	}
@@ -444,7 +441,7 @@ class UpdraftPlusAddons2 {
 	 * @param  string $checkperiod Period to check
 	 * @return boolean
 	 */
-	public function puc_check_now($shouldcheck, $lastcheck, $checkperiod) {// phpcs:ignore Generic.CodeAnalysis.UnusedFunctionParameter.Found -- Filter use
+	public function puc_check_now($shouldcheck, $lastcheck, $checkperiod) {// phpcs:ignore VariableAnalysis.CodeAnalysis.VariableAnalysis.UnusedVariable -- Filter use
 	
 		// Skip checks immediately after a WP upgrade. This action has existed since WP 4.4. Since we're just trying to reduce server load spikes when WP core automatic security upgrades happen, that is adequate.
 		if (did_action('pre_auto_update')) return false;
@@ -773,7 +770,7 @@ class UpdraftPlusAddons2 {
 			'noadverts' => array(
 				'name' => 'Remove adverts',
 				'description' => 'Removes all adverts from the control panel and emails',
-				'shopurl' => '/shop/no-adverts/'
+				'shopurl' => '/shop/updraftplus-premium/'
 			),
 			'all' => array(
 				'name' => 'All addons',
@@ -783,37 +780,37 @@ class UpdraftPlusAddons2 {
 			'multisite' => array(
 				'name' => 'WordPress Network (multisite) support',
 				'description' => 'Adds support for WordPress Network (multisite) installations, allowing secure backup by the super-admin only',
-				'shopurl' => '/shop/network-multisite/'
+				'shopurl' => '/shop/updraftplus-premium/'
 			),
 			'fixtime' => array(
 				'name' => 'Fix Time',
 				'description' => 'Allows you to specify the exact time at which backups will run',
-				'shopurl' => '/shop/fix-time/'
+				'shopurl' => '/shop/updraftplus-premium/'
 			),
 			'morefiles' => array(
 				'name' => 'More Files',
 				'description' => 'Allows you to backup WordPress core, and other files in your web space',
-				'shopurl' => '/shop/more-files/'
+				'shopurl' => '/shop/updraftplus-premium/'
 			),
 			'sftp' => array(
 				'name' => 'SFTP and FTPS and SCP',
 				'description' => 'Allows SFTP and SCP as a cloud backup method, and encrypted FTP',
-				'shopurl' => '/shop/sftp/'
+				'shopurl' => '/shop/updraftplus-premium/'
 			),
 			'dropbox-folders' => array(
 				'name' => 'Dropbox Folders',
 				'description' => 'Allows you to organise your backups into Dropbox sub-folders',
-				'shopurl' => '/shop/dropbox-folders/'
+				'shopurl' => '/shop/updraftplus-premium/'
 			),
 			'morestorage' => array(
 				'name' => 'Multiple storage destinations',
 				'description' => 'Allows you to send a single backup to multiple destinations (e.g. Dropbox and Google Drive and Amazon)',
-				'shopurl' => '/shop/morestorage/'
+				'shopurl' => '/shop/updraftplus-premium/'
 			),
 			'webdav' => array(
 				'name' => 'WebDAV support',
 				'description' => 'Allows you to use the WebDAV and encrypted WebDAV protocols for remote backups',
-				'shopurl' => '/shop/webdav/'
+				'shopurl' => '/shop/updraftplus-premium/'
 			)
 		);
 	}
@@ -1014,7 +1011,7 @@ class UpdraftPlusAddons2 {
 			}
 			
 			if (403 == $code) {
-				$ip_addr = $updraftplus->get_outgoing_ip_address();
+				$ip_addr = $updraftplus->get_outgoing_ip_address(true);
 				if (false !== $ip_addr && false !== filter_var($ip_addr, FILTER_VALIDATE_IP)) {
 					$message .= '  IP: '.htmlspecialchars($ip_addr);
